+++ Engine/Graphics/Primitives/circle.py
@@ -20,6 +20,10 @@ class Circle(Base):
         self._cached_vertices = None
         self.window = window
         self.shader = None
+        if self.window.render_type == 0:
+            self.polygon = Polygon(self.vertexes, self.window)
+        else:
+            self.polygon = PolygonLegacy(self.vertexes)
     
     def calculateSize(self):
         if not self._dirty and self._cached_vertices is not None:
@@ -45,6 +49,8 @@ class Circle(Base):
         self._cached_vertices = self.vertexes.copy()
         self._dirty = False
 
+        self.polygon.setVertexes(self.vertexes)
+
     def drawCircle(self, mode:drawMode):
         
         # Optimization
@@ -61,19 +67,18 @@ class Circle(Base):
             self.shader._apply_uniforms()
         
         if self.window.render_type == 1:
-            polygon = PolygonLegacy(self.vertexes)
-            polygon.setColor(self.color)
-            polygon.setWidthLines(self.widthlines)
-            polygon.setPointSize(self.pointsize)
-            polygon.setTexCoords(self.uv)
-            polygon.setTexture(self.texture)
-            polygon.drawPolygon(mode)
-        elif self.window.render_type == 0:
-            polygon = Polygon(self.vertexes.copy(), self.window)
-            polygon.setColor(self.color)
-            polygon.setTexCoords(self.uv)
-            polygon.setTexture(self.texture)
-            polygon.drawPolygon(mode)
+            self.polygon.setColor(self.color)
+            self.polygon.setWidthLines(self.widthlines)
+            self.polygon.setPointSize(self.pointsize)
+            self.polygon.setTexCoords(self.uv)
+            self.polygon.setTexture(self.texture)
+            self.polygon.drawPolygon(mode)
+        else:
+            self.polygon.vertexes = self.vertexes
+            self.polygon.setColor(self.color)
+            self.polygon.setTexCoords(self.uv)
+            self.polygon.setTexture(self.texture)
+            self.polygon.drawPolygon(mode)
         
         if self.shader:
             GL.glUseProgram(0)