+++ Engine/Graphics/Primitives/rectangle.py
@@ -18,7 +18,11 @@ class Rectangle(Base):
         self.calculated = False
         self.window = window
         self.shader = None
-    
+        if self.window.render_type == 0:
+            self.polygon = Polygon(self.vertexes, self.window)
+        else:
+            self.polygon = PolygonLegacy(self.vertexes)
+
     def calculateSize(self):
         center = Vec2(
             self.position.x + self.size.x / 2,
@@ -45,6 +49,8 @@ class Rectangle(Base):
             y_new = corners[i].x * sin_angle + corners[i].y * cos_angle
             self.vertexes[i] = Vec2(x_new + center.x, y_new + center.y)
 
+        self.polygon.setVertexes(self.vertexes)
+
         self.calculated = True
     
     def drawRectangle(self, mode:drawMode):
@@ -64,19 +70,18 @@ class Rectangle(Base):
             self.shader._apply_uniforms()
 
         if self.window.render_type == 1:
-            polygon = PolygonLegacy(self.vertexes)
-            polygon.setColor(self.color)
-            polygon.setWidthLines(self.widthlines)
-            polygon.setPointSize(self.pointsize)
-            polygon.setTexCoords(self.uv)
-            polygon.setTexture(self.texture)
-            polygon.drawPolygon(mode)
-        elif self.window.render_type == 0:
-            polygon = Polygon(self.vertexes.copy(), self.window)
-            polygon.setColor(self.color)
-            polygon.setTexCoords(self.uv)
-            polygon.setTexture(self.texture)
-            polygon.drawPolygon(mode)
+            self.polygon.setColor(self.color)
+            self.polygon.setWidthLines(self.widthlines)
+            self.polygon.setPointSize(self.pointsize)
+            self.polygon.setTexCoords(self.uv)
+            self.polygon.setTexture(self.texture)
+            self.polygon.drawPolygon(mode)
+        else:
+            self.polygon.vertexes = self.vertexes
+            self.polygon.setColor(self.color)
+            self.polygon.setTexCoords(self.uv)
+            self.polygon.setTexture(self.texture)
+            self.polygon.drawPolygon(mode)
         
         if self.shader:
             GL.glUseProgram(0)