+++ Engine/Graphics/__init__.py
@@ -3,4 +3,5 @@ from .Primitives.polygon import Polygon
 from .Primitives.rectangle import Rectangle
 from .Primitives.triangle import Triangle
 from .Primitives.circle import Circle
+from .Primitives.line import Line
 from .Particles.particles import simpleParticles